test: extract shared-suite policy and mark expensive problems slow
tests/test_mto2d.py imported PROBLEM_TEST_POLICIES from
tests/test_problem_implementations.py. That import pulls in
BUILTIN_PROBLEMS and therefore every problem's dependencies, so a missing
optional dependency failed a test that has nothing to do with it. Move
the policy into tests/problem_policies.py, which both modules import.

Drop artifacts_available and artifacts_reason: no problem ever set them,
and they were kept alive only by an assertion in the MTO2D test.

Add a `slow` marker, deselected by default via addopts, and apply it to
MTO2D's entry in the shared suite. Its simulation test pulls a 0.77 GB
image and runs a full 86,400-cell OpenFOAM case inside a CI job whose
30-minute budget covers the whole suite across four Python versions. Run
it explicitly with `pytest -m slow`.
