+++ pyproject.toml
@@ -222,6 +222,15 @@ docstring-code-line-length = "dynamic"
 
 
 ###################################### PYRIGHT ########################################
+[tool.pytest.ini_options]
+# `slow` covers problems whose smoke test pulls a large container image or runs a
+# long external simulation. They are deselected by default so the standard CI job
+# stays inside its time budget; run them explicitly with `pytest -m slow`.
+markers = [
+    "slow: pulls large container images or runs long simulations",
+]
+addopts = "-m 'not slow'"
+
 [tool.pyright]
 include = ["engibench/**"]
 exclude = ["**/node_modules", "**/__pycache__", "**/study*", "**/pyopt_history.py"]
